Adds new address mapping policies based on the ones in the XILINX FPGA of the project

+++ SpikeModel/src/utils.cpp
@@ -2,7 +2,7 @@
 
 namespace utils 
 {
-    uint8_t nextPowerOf2(uint64_t v)
+    uint64_t nextPowerOf2(uint64_t v)
     {
         v--;
         v |= v >> 1;
