+++ SpikeModel/src/utils.hpp
@@ -6,7 +6,7 @@
 
 namespace utils 
 {
-    uint8_t nextPowerOf2(uint64_t v);
+    uint64_t nextPowerOf2(uint64_t v);
 }
 #endif
